fix(libc): correct sqrtf, memcmp, fread/fwrite and 64-bit printf conversions
Four P1 defects in the freestanding libc, with host tests for the two that can
have them.

sqrtf was wrong across most of the exponent range. Newton-Raphson seeded with x
itself spends one iteration per octave before quadratic convergence begins, so
a fixed 12 iterations only covered arguments near 1. Measured against a double
reference: sqrtf(1e10) answered 2.4e6 for 1e5, sqrtf(3.4e38) was off by a factor
of 4.5e15, and -- not mentioned in the task entry -- small arguments were just as
broken, sqrtf(1e-20) being wrong by six orders. Seeding by halving the biased
exponent in the bit pattern (u >> 1, plus the 127<<22 the shift removes from the
bias) lands within a few percent for every finite x; four iterations then
converge. Worst relative error over 1e-20..3.4e38 falls from 4.5e15 to 6.8e-8,
inside a float ulp.

%lld/%llx truncated to 32 bits on wasm32: the whole formatting chain took long,
and vsnprintf cast long long down to it. utoa_base, format_unsigned and
format_signed now take long long, and the narrowing casts at the call sites are
gone.

fread reported a short read as EOF -- a console or pipe backend returns what it
has -- and dropped the bytes of a trailing partial item silently. fwrite
reported a short write only as a smaller item count, which a caller cannot tell
from having asked for fewer items. Both now loop to fill the request, advancing
only on positive progress so a backend answering 0 ends the loop instead of
spinning; fread sets eof only on a real 0-return and fwrite raises error on a
short write.

memcmp was the one entry point in string.c with no NULL guard. Its `lhs == rhs
|| count == 0` arms made NULL/NULL and zero-count safe by accident, so only a
single NULL with a non-zero count faulted.

Tests: new test_libc_math.c (sqrtf across the range, exact squares, domain
edges) and test_libc_string.c (memcmp guards and sign), both compiled against
the real sources. Each was checked against the OLD code and fails there:
test_sqrtf_large fails, test_libc_string segfaults.

-fno-builtin on both is load-bearing rather than hygiene. Written without it,
clang recognised sqrtf and constant-folded every literal call to the correct
IEEE result, so the cases never reached math.c and the suite produced identical
output against the broken and the fixed implementation -- it would have shipped
as coverage that tested nothing. The flag is commented at both call sites.

fread/fwrite and the printf widening have no test: a FILE here is backed by an
fd whose read/write go through FS IPC with no host stub, and the printf defect
is invisible on a host where long is 64 bits. Both gaps are filed rather than
papered over.

Gates: run-kernel-unit-tests green, quality.sh format --check and lint green,
run-qemu-test boots to CLI + calculator + halt, test_cli/test_hello_c/
test_hello_rust 18/18.

+++ src/libc/src/stdio.c
@@ -76,7 +76,9 @@ static void emit_string(stdio_emit_fn emit, void* ctx, const char* s, size_t len
     }
 }
 
-static size_t utoa_base(unsigned long value, unsigned int base, int uppercase, char* buffer,
+/* Widest integer the conversions handle. long is 32-bit on wasm32, so the whole
+ * chain is long long: narrowing here silently truncated every %lld/%llx. */
+static size_t utoa_base(unsigned long long value, unsigned int base, int uppercase, char* buffer,
                         size_t size) {
     static const char digits_lower[] = "0123456789abcdef";
     static const char digits_upper[] = "0123456789ABCDEF";
@@ -104,8 +106,8 @@ static size_t utoa_base(unsigned long value, unsigned int base, int uppercase, c
     return len;
 }
 
-static void format_unsigned(stdio_emit_fn emit, void* ctx, unsigned long value, unsigned int base,
-                            int uppercase, size_t width, char pad_char) {
+static void format_unsigned(stdio_emit_fn emit, void* ctx, unsigned long long value,
+                            unsigned int base, int uppercase, size_t width, char pad_char) {
     char digits[32];
     size_t len = utoa_base(value, base, uppercase, digits, sizeof(digits));
 
@@ -118,8 +120,9 @@ static void format_unsigned(stdio_emit_fn emit, void* ctx, unsigned long value, 
     emit_string(emit, ctx, digits, len);
 }
 
-static void format_signed(stdio_emit_fn emit, void* ctx, long value, size_t width, char pad_char) {
-    unsigned long magnitude = (unsigned long)value;
+static void format_signed(stdio_emit_fn emit, void* ctx, long long value, size_t width,
+                          char pad_char) {
+    unsigned long long magnitude = (unsigned long long)value;
     char digits[32];
     size_t len;
     size_t pad = 0;
@@ -127,7 +130,7 @@ static void format_signed(stdio_emit_fn emit, void* ctx, long value, size_t widt
 
     if (value < 0) {
         negative = 1;
-        magnitude = (unsigned long)(-(value + 1)) + 1ul;
+        magnitude = (unsigned long long)(-(value + 1)) + 1ull;
     }
 
     len = utoa_base(magnitude, 10u, 0, digits, sizeof(digits));
@@ -216,34 +219,33 @@ static int vformat(stdio_emit_fn emit, void* ctx, const char* format, va_list ar
         case 'd':
         case 'i':
             if (long_flag == 2) {
-                format_signed(emit, ctx, (long)va_arg(ap, long long), width, pad_char);
+                format_signed(emit, ctx, va_arg(ap, long long), width, pad_char);
             } else if (long_flag) {
                 format_signed(emit, ctx, va_arg(ap, long), width, pad_char);
             } else {
-                format_signed(emit, ctx, (long)va_arg(ap, int), width, pad_char);
+                format_signed(emit, ctx, (long long)va_arg(ap, int), width, pad_char);
             }
             break;
         case 'u':
             if (long_flag == 2) {
-                format_unsigned(emit, ctx, (unsigned long)va_arg(ap, unsigned long long), 10u, 0,
-                                width, pad_char);
+                format_unsigned(emit, ctx, va_arg(ap, unsigned long long), 10u, 0, width, pad_char);
             } else if (long_flag) {
                 format_unsigned(emit, ctx, va_arg(ap, unsigned long), 10u, 0, width, pad_char);
             } else {
-                format_unsigned(emit, ctx, (unsigned long)va_arg(ap, unsigned int), 10u, 0, width,
-                                pad_char);
+                format_unsigned(emit, ctx, (unsigned long long)va_arg(ap, unsigned int), 10u, 0,
+                                width, pad_char);
             }
             break;
         case 'x':
         case 'X':
             if (long_flag == 2) {
-                format_unsigned(emit, ctx, (unsigned long)va_arg(ap, unsigned long long), 16u,
-                                *format == 'X', width, pad_char);
+                format_unsigned(emit, ctx, va_arg(ap, unsigned long long), 16u, *format == 'X',
+                                width, pad_char);
             } else if (long_flag) {
                 format_unsigned(emit, ctx, va_arg(ap, unsigned long), 16u, *format == 'X', width,
                                 pad_char);
             } else {
-                format_unsigned(emit, ctx, (unsigned long)va_arg(ap, unsigned int), 16u,
+                format_unsigned(emit, ctx, (unsigned long long)va_arg(ap, unsigned int), 16u,
                                 *format == 'X', width, pad_char);
             }
             break;
